requirements.txt
# third party dependencies
#
# run 'pip install -r requirements.txt -t lib/' to install
Flask==0.10
nosegae
nose-exclude
